src/tasks/task8_final_presentation/requirements.txt
fastapi
tensorflow
uvicorn
numpy
pillow
python-multipart
streamlit

joblib
opencv-python-headless
pandas
scikit-image
scikit-learn
scikeras
gdown
torchvision
torch